# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cambridge Sustainable Food App API
  version: "0.1.0"
  description: >
    Canonical API contract for the CSF Food Flow Log. This file is the
    single source of truth for request/response shapes shared between
    csf-food-flow-api (Sanic/msgspec, hand-implemented against this spec)
    and csf-food-flow-pwa (Preact/TypeScript, types generated FROM this
    spec via `npm run generate:types`).

    Sync discipline: this file lives canonically in csf-food-flow-api/docs/
    and is copied into csf-food-flow-pwa/docs/ whenever it changes. There is
    no automated cross-repo sync -- if you change this spec, update both
    copies in the same change, and regenerate PWA types before committing.
servers:
  - url: /api/v1
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: csf_session
      description: >
        httpOnly, Secure, SameSite=None session cookie, set by
        /auth/magic-link/verify. Required cross-origin config: CORS_ORIGIN
        must be the PWA's exact origin (never "*") and requests must be
        made with credentials: 'include'.
  schemas:
    Role:
      type: string
      enum: [ADMIN, FOOD_CENTRE, HUB]
...
    LocationType:
      type: string
      enum: [HUB, FOOD_CENTRE]
...
    User:
      type: object
      required: [id, name, email, role, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        location_id: { type: [string, "null"], format: uuid }
        active: { type: boolean }
        last_login_at: { type: [string, "null"], format: date-time }
    Location:
      type: object
      required: [id, name, type, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: "#/components/schemas/LocationType" }
        active: { type: boolean }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    CategoryTotals:
      type: object
      properties:
        FRESH: { type: number }
        FROZEN: { type: number }
        AMBIENT: { type: number }
    WeeklyReport:
      type: object
      required: [week_start, week_end, totals, by_location]
      properties:
        week_start: { type: string, format: date }
        week_end: { type: string, format: date }
        totals:
          type: object
          required: [in_by_category, out_by_category]
          properties:
            in_by_category: { $ref: "#/components/schemas/CategoryTotals" }
            out_by_category: { $ref: "#/components/schemas/CategoryTotals" }
        by_location:
          type: array
          items:
            type: object
            required: [location_id, location_name, in_by_category, out_by_category]
            properties:
              location_id: { type: string, format: uuid }
              location_name: { type: string }
              in_by_category: { $ref: "#/components/schemas/CategoryTotals" }
              out_by_category: { $ref: "#/components/schemas/CategoryTotals" }
  responses:
    ValidationError:
      description: Input failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated but not permitted for this role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
paths:
...
  /locations:
    get:
      operationId: listLocations
      summary: List locations according to the authenticated user's role
      description: >
        HUB users receive only their own assigned location, never the full
        list. ADMIN may include inactive locations via `active=false`.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/LocationType" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Location" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createLocation
      summary: Create a new hub or food centre
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type: { $ref: "#/components/schemas/LocationType" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /locations/{id}:
    patch:
      operationId: updateLocation
      summary: Update a location name, type, or active status
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                type: { $ref: "#/components/schemas/LocationType" }
                active: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Not found }
...
  /reports/weekly:
    get:
      operationId: getWeeklyReport
      summary: Retrieve Monday-to-Sunday weekly totals by location and category
      description: >
        HUB may only view their own hub's report; any other requested
        location_id is ignored or rejected server-side. week_start must be
        a Monday or the API returns a validation error.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: week_start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: location_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeeklyReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: week_start is not a Monday
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /reports/weekly/export.csv:
    get:
      operationId: exportWeeklyCsv
      summary: >
        Export the Monday-to-Sunday weekly report matching the current
        spreadsheet layout
      description: >
        V1 restricts this to ADMIN. Exact column layout is PENDING
        operational validation against the real CSF spreadsheet template --
        see Step 3/4 notes.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: week_start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: location_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users:
    get:
      operationId: listUsers
      summary: List users for administration
      security: [{ cookieAuth: [] }]
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/Role" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createUser
      summary: Create a new user account without a password
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, role]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
                location_id: { type: [string, "null"], format: uuid }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /users/{id}:
    patch:
      operationId: updateUser
      summary: Update user name, email, role, location, or active status
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
                location_id: { type: [string, "null"], format: uuid }
                active: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Not found }